pyproject.toml
[project]
name = "closet-muse"
version = "0.1.0"
description = "AI based fashion assistant to help you style your wardrobe"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.120.4",
    "google-genai>=1.46.0",
    "ipykernel>=7.0.1",
    "pillow==11.3.0",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.11.0",
    "python-dotenv>=1.2.0",
    "sqlmodel>=0.0.27",
    "streamlit==1.50.0",
    "streamlit-image-select>=0.6.0",
    "uvicorn>=0.38.0",
]
